Move BpmnVisu out of the graph directory
